+++ src/components/Lista.module.css
@@ -1,3 +1,10 @@
+.tableContainer {
+  position: absolute;
+  top: 385px;
+  width: 100%;
+  overflow: auto;
+}
+
 h1 {
   background-color: DarkBlue;
   color: white;
@@ -8,7 +15,8 @@ h1 {
 .table {
   width: 100%;
   border: 3px solid #aabfbf;
-  height: auto;
+  max-width: 100%;
+  flex: 1;
 }
 
 .td {
@@ -24,12 +32,6 @@ h1 {
   height: 30px;
   background-color: CornflowerBlue;
   color: white;
-}
-
-.tableContainer {
-  max-height: 100%;
-  overflow: auto;
-
 }
 
 .icon {
