+++ src/components/Search.module.css
@@ -7,10 +7,42 @@
     margin: 5px 5px 8px 5px;
     background-color: #fff;
     height: 210px;
+    width: 70%;
 }
 
 .term {
     display: flex;
     flex-direction: column;
     margin: 30px;
+    display: flex;
+    align-items: center;
+}
+
+.term select {
+    display: flex;
+    justify-content: center;
+}
+
+@media screen and (max-width: 1200px) {
+    .search {
+        display: flex;
+        flex-direction: row;
+        width: 100%;
+        height: 100px;
+    }
+
+    .search div {
+        display: flex;
+        flex-direction: row;
+        width: 100%;
+    }
+
+    .term {
+        margin: 10px;
+        width: 100%;
+    }
+
+    .term select {
+        width: auto;
+    }
 }
