+++ src/context/ContatosProvider.js
@@ -1,17 +1,16 @@
 import React, { useState, useMemo, useEffect } from 'react';
 import { node } from 'prop-types';
 import ContatosContext from './ContatosContext';
-// import { act } from 'react-dom/test-utils';
 import { contatosData } from '../mocks/contatosData';
-// import { act } from 'react-dom/test-utils';
+import { act } from 'react-dom/test-utils';
 
 function ContatosProvider({ children }) {
   // API
   const [loading, setLoading] = useState(true);
-  const [listaContatos, setListaContatos] = useState([]); // carrega as informações da mock
+  const [listaContatos, setListaContatos] = useState([]);
 
   // New
-  const [newContato, setNewContato] = useState(false); //abre modal new
+  const [newContato, setNewContato] = useState(false);
   const [formData, setFormData] = useState({
     first_name: '',
     last_name: '',
@@ -20,12 +19,12 @@ function ContatosProvider({ children }) {
     language: '',
     avatar: '',
     birthday: '',
-  }); // salva as alterações do form new
+  });
 
   // Edit
-  const [openEditContato, setOpenEditContato] = useState(false); // abre modal edit
-  const [idContatoEdit, setIdContatoEdit] = useState(0); // seleciona contato para editar
-  const [contatoEdited, setContatoEdited] = useState(); // salva as alterações do form de edição
+  const [openEditContato, setOpenEditContato] = useState(false); 
+  const [idContatoEdit, setIdContatoEdit] = useState(0);
+  const [contatoEdited, setContatoEdited] = useState();
   
   // Search
   const [searchTerm, setSearchTerm] = useState({
@@ -33,10 +32,10 @@ function ContatosProvider({ children }) {
     language: '',
     age: '',
     birthday: '',
-  }); // pesquisa digitada
-  const [filtroContatos, setFiltroContatos] = useState(''); // listaContatos === searchTerm
-  const [birthday, setBirthday] = useState('') // carregas as datas de aniversario
-  const [language, setLanguage] = useState(''); // carrega as languages
+  });
+  const [filtroContatos, setFiltroContatos] = useState('');
+  const [birthday, setBirthday] = useState('')
+  const [language, setLanguage] = useState('');
 
   //Gráficos
   const [languageTotals, setLanguageTotals] = useState({});
@@ -50,6 +49,7 @@ function ContatosProvider({ children }) {
       } else {
         const response = await fetch('https://my.api.mockaroo.com/lestetelecom/test.json?key=f55c4060');
         if (!response.ok) {
+          act(() => { setLoading(false); });
           return setListaContatos(contatosData);
         }
         const data = await response.json();
